requirements.txt
google-cloud-bigquery
python-dotenv
google-generativeai
google-adk
fastapi
uvicorn
google-cloud-resource-manager
sse-starlette 